cos/token_bucket.rs: colocate 2 lease-top-up tests with production (#984 P3 Phase 4a)
Move maybe_top_up_cos_queue_lease_unblocks_local_exact_queue_without_tokens
and maybe_top_up_cos_root_lease_unblocks_large_frame_exceeding_lease_bytes
from tx/mod.rs::tests to a new mod tests in cos/token_bucket.rs.

New tests block imports CoSQueueConfig and the queue_service helpers
select_cos_guarantee_batch / _with_fast_path that the tests exercise
to validate the token-bucket lease top-up reactivates queues that
were previously starved.

cargo test --bins 865/0/2.
